Move to gpt2 file so it gets tested
